// src/shared/api/models/postApiSaleBodyDataItemsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiSaleBodyDataItemsItemProductId } from './postApiSaleBodyDataItemsItemProductId';
import type { PostApiSaleBodyDataItemsItemBundleOfferId } from './postApiSaleBodyDataItemsItemBundleOfferId';
import type { PostApiSaleBodyDataItemsItemServiceId } from './postApiSaleBodyDataItemsItemServiceId';
import type { PostApiSaleBodyDataItemsItemEmployeeId } from './postApiSaleBodyDataItemsItemEmployeeId';
import type { PostApiSaleBodyDataItemsItemItemType } from './postApiSaleBodyDataItemsItemItemType';
import type { PostApiSaleBodyDataItemsItemSaleType } from './postApiSaleBodyDataItemsItemSaleType';
import type { PostApiSaleBodyDataItemsItemItemImageUrl } from './postApiSaleBodyDataItemsItemItemImageUrl';
import type { PostApiSaleBodyDataItemsItemUnitId } from './postApiSaleBodyDataItemsItemUnitId';
import type { PostApiSaleBodyDataItemsItemUnitName } from './postApiSaleBodyDataItemsItemUnitName';
import type { PostApiSaleBodyDataItemsItemUnitAbbreviation } from './postApiSaleBodyDataItemsItemUnitAbbreviation';
import type { PostApiSaleBodyDataItemsItemCategoryId } from './postApiSaleBodyDataItemsItemCategoryId';
import type { PostApiSaleBodyDataItemsItemDiscountType } from './postApiSaleBodyDataItemsItemDiscountType';
import type { PostApiSaleBodyDataItemsItemTaxesItem } from './postApiSaleBodyDataItemsItemTaxesItem';
import type { PostApiSaleBodyDataItemsItemGiftCard } from './postApiSaleBodyDataItemsItemGiftCard';
import type { PostApiSaleBodyDataItemsItemDiscountsItem } from './postApiSaleBodyDataItemsItemDiscountsItem';

export type PostApiSaleBodyDataItemsItem = {
    productId?: PostApiSaleBodyDataItemsItemProductId;
    bundleOfferId?: PostApiSaleBodyDataItemsItemBundleOfferId;
    serviceId?: PostApiSaleBodyDataItemsItemServiceId;
    employeeId?: PostApiSaleBodyDataItemsItemEmployeeId;
    itemName?: string;
    itemType?: PostApiSaleBodyDataItemsItemItemType;
    saleType?: PostApiSaleBodyDataItemsItemSaleType;
    quantity: number;
    amount: number;
    discountAmount?: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    taxAmount?: number;
    price: number;
    note?: string;
    description?: string;
    itemImageUrl?: PostApiSaleBodyDataItemsItemItemImageUrl;
    unitId?: PostApiSaleBodyDataItemsItemUnitId;
    unitName?: PostApiSaleBodyDataItemsItemUnitName;
    unitAbbreviation?: PostApiSaleBodyDataItemsItemUnitAbbreviation;
    categoryId?: PostApiSaleBodyDataItemsItemCategoryId;
    categoryName?: string;
    discountType?: PostApiSaleBodyDataItemsItemDiscountType;
    taxes?: PostApiSaleBodyDataItemsItemTaxesItem[];
    giftCard?: PostApiSaleBodyDataItemsItemGiftCard;
    discounts?: PostApiSaleBodyDataItemsItemDiscountsItem[];
};